requirements.txt
crewai[tools]
litellm
python-dotenv
flask
flask-sqlalchemy
flask-login
psycopg2-binary
gunicorn
pandas
reportlab
pytest
